Add metadata (file, namespace, line numbers)
Getting metadata about the file, namespace and line the log function was
called from requires the use of macros.

So convert the log functions users would call (log, spy, spy-with) to
macros to get this information. The output functions have also been
modified to include the namespace and line number (and filename, for
json output) in the default case.

A slight downside to this is that the end-to-end tests include the line
number in the output, so adding or removing lines from the test file
causes those tests to fail... Any better ideas for how to test this much
appreciated.