Add thread-first versions of the spy functions
spy and spy-with were specifically thread-last due to the argument
order. This renames them to spy->> and spy-with->> to make that
explicit, and also adds a thread-first variant of each.

Also updated the doc for each case to actually be correct, i.e. called
in a thread-first or -last way, with data strctures that are manipulated
in thread-first or -last context.